requirements.txt
pyodbc
openai
pinecone
langchain
langchain-community
python-dotenv
beautifulsoup4
tiktoken
langchain-pinecone